python/examples/by-use-case/solana-purchase-on-amazon/requirements.txt
python-dotenv>=1.0.0
openai>=1.7.0
solana>=0.34.0
solders>=0.21.0
base58>=2.1.0
pydantic>=2.0.0
aiohttp>=3.9.5
goat-sdk>=0.2.0
goat-sdk-wallet-solana>=0.2.0
goat-sdk-plugin-crossmint-headless-checkout>=0.1.0

# Note: readline and asyncio are part of the Python standard library so they don't need installation

# For local development, install goat packages from local directory:
# pip install -e ../../../src/goat-sdk
# pip install -e ../../../src/wallets/solana
# pip install -e ../../../src/plugins/headless 